Update LiFi.ino
Eliminated the need for  WriteChar function. Replaced 'int' with 'byte' wherever necessary. Did away with the need for passing 7 arguments to the LightFlash function, replacing with only 1 byte argument. Did away with the requirement of switch-case, for most portion of the code. Tried to focus more upon the inherent properties of the loop function, which is that it is called infinitely